DONE - BUG - Door sounds are not right - added new effect - check for below 0 DONE - BUG - Some script functions are not being found when called from console DONE - All droids weapon recharge based off base setting for 001 DONE - Fix up player firing rate - needs to be slower DONE - Lock database view to current droid DONE - Put in counter to stop continual direction swapping - limit to 3 then ignore DONE - What to do when there is no path for flee - default to change back to PATROL DONE - Using lookup table for particle explosion DONE - Stop lowEnergy playing when entering transfer mode DONE - Droid patrol AI - use direction counter more - try to find another waypoint if pathing fails DONE - Healing tiles now animated in deck view DONE - Reworked logic around collisions and bullets DONE - BUG - If on low health, color is wrong for player sprite in transfer screen DONE - BUG - Player bullets bouncing off droid when in collision - was droid to droid bullets DONE - BUG - Droids bullets are bouncing off each other, not being destroyed DONE - BUG - Player fade in low health disappears from view - alpha setting wrong - from bump damage? DONE - BUG - Highscore not working - always adding name when score is too low

+++ hdr/game/database.h
@@ -14,7 +14,7 @@ typedef struct
 	int         chanceToShoot;
 	int         bulletDamage;
 	bool        disrupterImmune;
-	float       rechargeTime;
+	double      rechargeTime;
 	int         tokenCount;
 	int         imageIndex;
 	std::string dbImageFileName;
